# perf(python): Improve tuple and list serializer performance