@penguin-wwy 2024-11-07
> This is great! We have a new format which will improve performance a lot: https://fury.apache.org/docs/specification/fury_xlang_serialization_spec#list. Would you like to implement this format for pyfury?

Okay, I will implement it.